requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Database connectivity
psycopg2-binary==2.9.9
pymongo==4.6.1
sqlalchemy==2.0.23

# ML Libraries
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2

# Survival Analysis
lifelines==0.27.8

# Feature Engineering
category_encoders==2.6.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# ML Ops
mlflow==2.9.2
optuna==3.5.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
jupyter==1.0.0
ipykernel==6.27.1

# Data Quality
great-expectations==0.18.8

# Scheduling (optional)
schedule==1.2.0

# API (для будущего)
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
